Add dependency check before deleting cities, add set-on-insert for some response fields in HandymanProfileDetailsResponse.

+++ HandyLink.Model/Responses/HandymanProfileDetailsResponse.cs
@@ -17,7 +17,7 @@ namespace HandyLink.Model.Responses
         public List<ReviewListResponse> Reviews { get; set; } = new();
         public int JobsCompleted { get; set; } //set in service
         public int ExperienceYears { get; set; }//set in service
-        public int AverageRating { get; set; }//set in service
+        public double AverageRating { get; set; }//set in service
         public int ReviewsCount { get; set; } //set in service
     }
 }
